requirements.txt
streamlit==1.28.0
yfinance==0.2.28
numpy==2.1.1
pandas==2.2.3
plotly==5.15.0
statsmodels==0.14.0
requests==2.31.0
python-dotenv==1.0.0